🐛 fix authOptions import error
