+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,51 +1,11 @@
-import NextAuth, { type DefaultSession, type NextAuthOptions } from 'next-auth'
-import FacebookProvider from 'next-auth/providers/facebook'
-import GoogleProvider from 'next-auth/providers/google'
-import TwitterProvider from 'next-auth/providers/twitter'
+import { authOptions } from '@/lib/auth/authOptions'
+import NextAuth, { type DefaultSession } from 'next-auth'
 
 declare module 'next-auth' {
   interface Session {
     user: {
       id: string
     } & DefaultSession['user']
-  }
-}
-
-export const authOptions: NextAuthOptions = {
-  providers: [
-    GoogleProvider({
-      clientId: process.env.GOOGLE_CLIENT_ID ?? '',
-      clientSecret: process.env.GOOGLE_CLIENT_SECRET ?? ''
-    }),
-    // TODO: Add Facebook provider
-    FacebookProvider({
-      clientId: process.env.FACEBOOK_CLIENT_ID ?? '',
-      clientSecret: process.env.FACEBOOK_CLIENT_SECRET ?? ''
-    }),
-    // TODO: Add Twitter provider
-    TwitterProvider({
-      clientId: process.env.TWITTER_CLIENT_ID ?? '',
-      clientSecret: process.env.TWITTER_CLIENT_SECRET ?? '',
-      version: '2.0'
-    })
-  ],
-  callbacks: {
-    async jwt({ token, user }) {
-      if (user?.id) {
-        token.id = user.id
-      }
-      return token
-    },
-    async session({ session, token }) {
-      if (session.user) {
-        session.user.id = token.id as string
-      }
-      return session
-    }
-  },
-  pages: {
-    signIn: '/auth/signin',
-    error: '/auth/error'
   }
 }
